[Core] Added separate dev cog
This is enabled with the cli flag --dev
