--- core/__init__.py
@@ -1,5 +1,0 @@
-from .owner import Owner
-
-
-def setup(bot):
-    bot.add_cog(Owner())
