+++ core/utils/chat_formatting.py
@@ -44,14 +44,14 @@ def pagify(text, delims=["\n"], *, escape=True, shorten_by=8,
                              for d in delims])
         closest_delim = closest_delim if closest_delim != -1 else page_length
         if escape:
-            to_send = escape_mass_mentions(in_text[:closest_delim])
+            to_send = escape(in_text[:closest_delim], mass_mentions=True)
         else:
             to_send = in_text[:closest_delim]
         yield to_send
         in_text = in_text[closest_delim:]
 
     if escape:
-        yield escape_mass_mentions(in_text)
+        yield escape(in_text, mass_mentions=True)
     else:
         yield in_text
